速度环调成（电池电压11.6V下，高速可跑2m直线）

+++ Hardware/control.h
@@ -8,7 +8,7 @@
 /* å…?è·?å¾?è¿? */
 extern int track1, track2, track3, track4, track5, track6;
 extern int left_speed, right_speed;
-extern volatile uint8_t target_speed;
+extern volatile int target_speed;
 extern int left_encoder_speed;
 extern int right_encoder_speed;
